Fix linux build (hack)
We need to import this commit :

https://github.com/kubernetes/kubernetes/commit/49201f6923b2a7ba6694ef1ed8db6f9a24e7eb2e

To get the linux build compiled.

Currently the glide dependency is a mess and we can't get that commit in openshift/kubernetes
without some awfull circular dependencies :sadpanda:

For a quick fix we import the patch directly on the sources.

Proper solution would be :

        - Cleanup the depencies.
        - Work with upstream on an openshift origin client like for kube.
        - Restart to a minimal and sane glide.yaml file letting glide do its magic.