Optimize Makefile
This has been a bit of a pet of mine to get the compile time.

By default we run go fmt which can be awfully slow and then change
the timestamp so make(1) cannot detect if there is a change or not and
alway recompile[1]

It does run by default the unittests on every make command which is
something that takes quite some time as well.

Let's move those in the standard `test` target making sure this is
available in make release which arguably is the best place to have
those.

[1] which break my editor on the fly compilation for me.